requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
kafka-python
redis
